+++ config/response_builder.php
@@ -108,6 +108,35 @@ return [
 
 	/*
 	|--------------------------------------------------------------------------
+	| Response JSON keys mapping
+	|--------------------------------------------------------------------------
+	|
+	| You can remap response JSON keys if really needed.
+	|
+	|
+	| WARNING: there's NO duplicate check at runtime, so if you remap two keys
+	| to the same values you will end up with problems. There's testing trait
+	| to prevent this from happening, so ensure you unit test your app (see docs!)
+	|
+	| NOTE: Ensure you have this config file using:
+	|
+	|    use MarcinOrlowski\ResponseBuilder\ResponseBuilder;
+	|
+	| if you want to use custom mapping.
+	|
+	| It's safe to completely remove/comment out this config element.
+	|
+	*/
+//	'response_key_map' => [
+//		ResponseBuilder::KEY_SUCCESS => 'success',
+//		ResponseBuilder::KEY_CODE    => 'code',
+//		ResponseBuilder::KEY_LOCALE  => 'locale',
+//		ResponseBuilder::KEY_MESSAGE => 'message',
+//		ResponseBuilder::KEY_DATA    => 'data',
+//	],
+
+	/*
+	|--------------------------------------------------------------------------
 	| Exception handler error codes
 	|--------------------------------------------------------------------------
 	|